+++ ErsatzTV.Infrastructure.MySql/Migrations/20251201160022_Add_RemoteStreamMetadata_ContentRatingPlot.cs
@@ -1,0 +1,40 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace ErsatzTV.Infrastructure.MySql.Migrations
+{
+    /// <inheritdoc />
+    public partial class Add_RemoteStreamMetadata_ContentRatingPlot : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<string>(
+                name: "ContentRating",
+                table: "RemoteStreamMetadata",
+                type: "longtext",
+                nullable: true)
+                .Annotation("MySql:CharSet", "utf8mb4");
+
+            migrationBuilder.AddColumn<string>(
+                name: "Plot",
+                table: "RemoteStreamMetadata",
+                type: "longtext",
+                nullable: true)
+                .Annotation("MySql:CharSet", "utf8mb4");
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "ContentRating",
+                table: "RemoteStreamMetadata");
+
+            migrationBuilder.DropColumn(
+                name: "Plot",
+                table: "RemoteStreamMetadata");
+        }
+    }
+}