+++ ErsatzTV.Infrastructure.Sqlite/Migrations/20251201155729_Add_RemoteStreamMetadata_ContentRatingPlot.cs
@@ -1,0 +1,38 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace ErsatzTV.Infrastructure.Sqlite.Migrations
+{
+    /// <inheritdoc />
+    public partial class Add_RemoteStreamMetadata_ContentRatingPlot : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<string>(
+                name: "ContentRating",
+                table: "RemoteStreamMetadata",
+                type: "TEXT",
+                nullable: true);
+
+            migrationBuilder.AddColumn<string>(
+                name: "Plot",
+                table: "RemoteStreamMetadata",
+                type: "TEXT",
+                nullable: true);
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "ContentRating",
+                table: "RemoteStreamMetadata");
+
+            migrationBuilder.DropColumn(
+                name: "Plot",
+                table: "RemoteStreamMetadata");
+        }
+    }
+}